Move navlink and megamenu to components (#478)
* Move NavLink to components

* Move MegaDropdownNavbarItem to components
