+++ src/components/MegaDropdownNavbarItem/index.tsx
@@ -14,10 +14,10 @@ import {
   Collapsible,
   useLocalPathname,
 } from '@docusaurus/theme-common';
-import { NavLink } from './DefaultNavbarItem';
+import NavLink from '@site/src/components/NavLink';
 import NavbarItem from '@theme/NavbarItem';
-import './styles.css';
 import type { Props as NavLinkProps } from '@theme/NavbarItem/DefaultNavbarItem';
+import './styles.css';
 
 const dropdownLinkActiveClass = 'dropdown__link--active';
