+++ src/components/MegaDropdownNavbarItem/styles.css
@@ -1,0 +1,17 @@
+.dropdown--mega {
+  position: static;
+}
+
+.dropdown--mega .dropdown__menu {
+  visibility: inherit;
+  border-radius: 0;
+  opacity: 1;
+  box-shadow: none;
+  width: 100%;
+  position: static;
+  transform: none;
+  padding: var(--ifm-navbar-padding-vertical)
+    var(--ifm-navbar-padding-horizontal);
+  border-bottom: var(--ifm-global-border-width) solid
+    var(--ifm-color-emphasis-200);
+}
